% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{animation_raster}
\alias{animation_raster}
\title{Animate SpatRast as GIF}
\usage{
animation_raster(data, AOI = NULL, outfile, colors = blues9)
}
\arguments{
\item{data}{a single SpatRast object}

\item{AOI}{optional AOI sf or SpatVect object to overlay on gif}

\item{outfile}{path to write gif file, must have .gif extenstion}

\item{colors}{colors to plot with}
}
\value{
file.path
}
\description{
Animate a SpatRaster object as a gif.
}
\seealso{
Other viz: 
\code{\link{animation_vector}()},
\code{\link{animation}()}
}
\concept{viz}
