% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{animation_vector}
\alias{animation_vector}
\title{Animate vector as GIF}
\usage{
animation_vector(data, feild_pattern = NULL, outfile, colors = blues9)
}
\arguments{
\item{data}{a SpatVect or sf object}

\item{feild_pattern}{optional string vector to filter the desired attributes by}

\item{outfile}{path to write gif file, must have .gif extenstion}

\item{colors}{colors to plot with}
}
\value{
file.path
}
\description{
Animate a sf or SpatVect object as a gif.
}
\seealso{
Other viz: 
\code{\link{animation}()},
\code{\link{animation_raster}()}
}
\concept{viz}
