% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_define_dates.R
\name{build.date}
\alias{build.date}
\title{string to start and end dates}
\usage{
build.date(startDate, endDate)
}
\arguments{
\item{startDate}{(character) starting date}

\item{startDate}{(character) ending date}
}
\value{
the number of days in current month
}
\description{
**INTERNAL** convert strings to start and end dates
}
\keyword{internal}
