% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonalStats.R
\name{buildWeights}
\alias{buildWeights}
\title{Build Area Weighted Intersection Data Frame}
\usage{
buildWeights(grid, poly, abb = NULL)
}
\arguments{
\item{grid}{a sf data.frame}

\item{poly}{a sf data.frame}

\item{abb}{The feature (poly) attribute to intersect by}
}
\value{
a sf layer of polygons
}
\description{
Returns the fractional percent of each grid feature covered by each intersecting polygon feature.
These can be used as the weights in an area-weighted mean overlay analysis zonal
}
\examples{
\dontrun{
AOI = getAOI(state = "CA", county = "all")

r = getMACA(AOI, "prcp", startDate = "2018-10-10")
g = makeGrid(r)
w = buildWeights(g, AOI, "countyfp")
}
}
\author{
Mike Johnson
}
