% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateContractionConvergence.R
\name{calculateContractionConvergence}
\alias{calculateContractionConvergence}
\title{Calculate budgets for the Contraction / Convergence approach}
\usage{
calculateContractionConvergence(data, Ct, global_emissions, population_perc)
}
\arguments{
\item{data}{A dataframe column or list which is going to contain the emission budgets. Should already be established and contain the country's base year emissions.}

\item{Ct}{Convergence factor for every year: Ct = (Year - BaseYear) / (ConvergenceYear -  BaseYear)}

\item{global_emissions}{Path for global emissions for every year}

\item{population_perc}{Percentage of every country's population on global population for every year}
}
\description{
Calculate budgets for the Contraction / Convergence approach
}
\details{
Calculates emission budgets for every country and every year
}
