% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\docType{data}
\name{catalog}
\alias{catalog}
\title{ClimateR Catalog}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 107857 rows and 28 columns.
}
\usage{
catalog
}
\description{
ClimateR Catalog
}
\seealso{
Other catalog: 
\code{\link{read_live_catalog}()}
}
\concept{catalog}
\keyword{datasets}
