% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netrc_utils.R
\name{checkDodsrc}
\alias{checkDodsrc}
\title{Check dodsrc file}
\usage{
checkDodsrc(dodsrcFile = getDodsrcPath(), netrcFile = getNetrcPath())
}
\arguments{
\item{dodsrcFile}{File path to dodsrc file to check.}

\item{netrcFile}{File path to netrc file to check.}
}
\value{
logical
}
\description{
Check that there is a netrc file with a valid entry for urs.earthdata.nasa.gov.
}
\seealso{
Other netrc: 
\code{\link{checkNetrc}()},
\code{\link{getDodsrcPath}()},
\code{\link{getNetrcPath}()},
\code{\link{writeDodsrc}()},
\code{\link{writeNetrc}()}
}
\concept{netrc}
