% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_climateR.R
\docType{package}
\name{climateR}
\alias{climateR}
\alias{climateR-package}
\title{climateR}
\description{
\code{climateR} package
}
\details{
Download gridded and point climate data

See the README on
}
