% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{climater_dap}
\alias{climater_dap}
\title{ClimateR dry run}
\usage{
climater_dap(id, args, verbose, dryrun, print.arg = FALSE)
}
\arguments{
\item{id}{The resource name, agency, or catalog identifier}

\item{args}{The parent function arguments}

\item{verbose}{Should messages be emited?}

\item{dryrun}{Return summary of data prior to retrieving it}

\item{print.arg}{should arguments be printed? Usefull for debugging}
}
\value{
data.frame
}
\description{
ClimateR dry run
}
\seealso{
Other dap: 
\code{\link{.resource_grid}()},
\code{\link{.resource_time}()},
\code{\link{climater_filter}()},
\code{\link{dap}()},
\code{\link{dap_crop}()},
\code{\link{dap_get}()},
\code{\link{dap_meta}()},
\code{\link{dap_summary}()},
\code{\link{dap_to_local}()},
\code{\link{dap_xyzv}()},
\code{\link{extract_sites}()},
\code{\link{get_data}()},
\code{\link{go_get_dap_data}()},
\code{\link{grid_meta}()},
\code{\link{make_ext}()},
\code{\link{make_vect}()},
\code{\link{merge_across_time}()},
\code{\link{parse_date}()},
\code{\link{read_dap_file}()},
\code{\link{read_ftp}()},
\code{\link{time_meta}()},
\code{\link{try_att}()},
\code{\link{var_to_terra}()},
\code{\link{variable_meta}()},
\code{\link{vrt_crop_get}()}
}
\concept{dap}
