% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBarChartExemplaryStates.R
\name{createBarChartExemplaryStates}
\alias{createBarChartExemplaryStates}
\title{Bar chart: Budget changes depending on justice approach for exemplary states}
\usage{
createBarChartExemplaryStates(data, theme, cols, selected_budget)
}
\arguments{
\item{data}{Data frame containing the data}

\item{theme}{Some ggplot theme adjustments}

\item{cols}{The color values for scale_color_manual}
}
\description{
Bar chart: Budget changes depending on justice approach for exemplary states
}
