% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBarChartLeadersLaggards.R
\name{createBarChartLeadersLaggards}
\alias{createBarChartLeadersLaggards}
\title{Bar chart: Leaders and laggard countries regarding emission budget}
\usage{
createBarChartLeadersLaggards(data, scales, theme)
}
\arguments{
\item{data}{Data frame containing the data}
}
\description{
Bar chart: Leaders and laggard countries regarding emission budget
}
\details{
Creates the bar chart
}
