% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBarChartYearsLeft.R
\name{createBarChartYearsLeft}
\alias{createBarChartYearsLeft}
\title{Bar chart: Years left per country}
\usage{
createBarChartYearsLeft(data, base_year, theme, cols)
}
\arguments{
\item{data}{Data frame with data to plot}

\item{base_year}{Numeric: Which base year to select?}

\item{theme}{Some ggplot theme adjustments}

\item{cols}{The colors for the points}
}
\description{
Bar chart: Years left per country
}
