% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createChartScales.R
\name{createChartScales}
\alias{createChartScales}
\title{Create Chart scales}
\usage{
createChartScales(data, props)
}
\arguments{
\item{data}{Atomic numeric vector}

\item{props}{Probabilities for quantiles}
}
\description{
Create Chart scales
}
\details{
Creates numbers used for a color scale in charts. Seven colors are used, center is zero.
}
