% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createChloroplethChart.R
\name{createChloroplethChart}
\alias{createChloroplethChart}
\title{Chloropleth Chart}
\usage{
createChloroplethChart(data, palette, tooltip, data_column, chart_title)
}
\arguments{
\item{data}{SpatialPolygonsDataFrame containing the emission data to be plotted nested within "data"}

\item{palette}{Color palette, produced e.g. by means of colorNumeric()}

\item{tooltip}{Character vector containing tooltips for each data row}

\item{data_column}{Name of the column which contains the value which should be plotted (one value per country)}

\item{title}{The title of the plot}
}
\description{
Chloropleth Chart
}
\details{
Creates a chloropleth chart by means of the leaflet package
}
