% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHeatmapEmissionBudgetLeft.R
\name{createHeatmapEmissionBudgetLeft}
\alias{createHeatmapEmissionBudgetLeft}
\title{Heatmap: Budget left for all years}
\usage{
createHeatmapEmissionBudgetLeft(data, scales, from, to, theme)
}
\arguments{
\item{data}{Data frame containing the calculated "just" emissions budget for each state. Needs to have columns "country" and "budget_left_perc"}

\item{scales}{Values for color scale, with seven vakues}

\item{from}{Start year}

\item{to}{Final year}
}
\description{
Heatmap: Budget left for all years
}
\details{
Prepares the heatmap displaying the budget which is left for all countries and years
}
