% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRectPlotEmissionsRegion.R
\name{createRectPlotEmissionsRegion}
\alias{createRectPlotEmissionsRegion}
\title{Rect plot: Emissions per region and capita}
\usage{
createRectPlotEmissionsRegion(data, theme, cols, year, pos_label)
}
\arguments{
\item{data}{Data frame containing the emission data}
}
\description{
Rect plot: Emissions per region and capita
}
\details{
Prepares the rect plot
}
