% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRibbonChartScenarios.R
\name{createRibbonChartScenarios}
\alias{createRibbonChartScenarios}
\title{Ribbon chart: Future emission scenarios}
\usage{
createRibbonChartScenarios(data_ribbon, data_column, theme)
}
\arguments{
\item{data_ribbon}{Data frame with data for ribbons}

\item{data_column}{Data frame with data for the column on the right}

\item{theme}{Some ggplot theme adjustments}
}
\description{
Ribbon chart: Future emission scenarios
}
