% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap.R
\name{dap_summary}
\alias{dap_summary}
\title{Print Summary Information About a OpenDAP Resource}
\usage{
dap_summary(dap = NULL, url = NULL)
}
\arguments{
\item{dap}{data.frame from catalog or dap_crop}

\item{url}{Unique Resource Identifier (http or local)}
}
\description{
Print summary information about a DAP summary
}
\seealso{
Other dap: 
\code{\link{.resource_grid}()},
\code{\link{.resource_time}()},
\code{\link{climater_dap}()},
\code{\link{climater_filter}()},
\code{\link{dap_crop}()},
\code{\link{dap_get}()},
\code{\link{dap_meta}()},
\code{\link{dap_to_local}()},
\code{\link{dap_xyzv}()},
\code{\link{dap}()},
\code{\link{extract_sites}()},
\code{\link{get_data}()},
\code{\link{go_get_dap_data}()},
\code{\link{grid_meta}()},
\code{\link{make_ext}()},
\code{\link{make_vect}()},
\code{\link{merge_across_time}()},
\code{\link{parse_date}()},
\code{\link{read_dap_file}()},
\code{\link{read_ftp}()},
\code{\link{time_meta}()},
\code{\link{try_att}()},
\code{\link{var_to_terra}()},
\code{\link{variable_meta}()},
\code{\link{vrt_crop_get}()}
}
\concept{dap}
