% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dap_to_local}
\alias{dap_to_local}
\title{Convert OpenDAP to start/count call}
\usage{
dap_to_local(dap, get = TRUE)
}
\arguments{
\item{dap}{dap description}

\item{get}{should data be collected?}
}
\value{
numeric array
}
\description{
Convert OpenDAP to start/count call
}
