% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dap_xyzv}
\alias{dap_xyzv}
\title{Get XYTV data from DAP URL}
\usage{
dap_xyzv(obj, varname = NULL, varmeta = FALSE)
}
\arguments{
\item{obj}{an OpenDap URL or NetCDF object}

\item{varmeta}{should variable metadata be appended?}
}
\value{
data.frame with (varname, X_name, Y_name, T_name)
}
\description{
Get XYTV data from DAP URL
}
