% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_define_dates.R
\name{define.dates}
\alias{define.dates}
\title{Define Date Matrix}
\usage{
define.dates(startDate, endDate, baseDate = NULL, type = "d")
}
\arguments{
\item{startDate}{character starting date for data retrieval in the form YYYY-MM-DD.}

\item{endDate}{character ending date for data retrieval in the form YYYY-MM-DD. Default is the startDate}

\item{baseDate}{Used to define the first data of a data set in the form YYYY-MM-DD. Used to align data indeces}
}
\value{
a data.frame with date, year, julien, and string representations of the data range
}
\description{
**INTERNAL** From a user supplied start and end date gnerate the julien, string, and date representations of the data used in climateR
}
\keyword{internal}
