% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_define_grid.R
\name{define.grid}
\alias{define.grid}
\title{Define AOI native grid}
\usage{
define.grid(AOI, source = NULL)
}
\arguments{
\item{AOI}{an POINT or POLYGON object}

\item{source}{a dataset id (source)  being called}
}
\value{
a list of values including type, lat.call, lon.call, extent, AOI, proj4string
}
\description{
**INTERNAL** This function defines the grid of the AOI in terms of the climate resource
}
\keyword{internal}
