% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_grid2.R
\name{define.grid2}
\alias{define.grid2}
\title{Define AOI native grid}
\usage{
define.grid2(AOI, url)
}
\arguments{
\item{AOI}{an AOI object}

\item{url}{a URL to the home OPENDaP resource}
}
\value{
a list of values including type, lat.call, lon.call, extent and AOI
}
\description{
This functions defines the grid of the AOI in terms of the climate resource
}
\author{
Mike Johnson
}
\keyword{internal}
