% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_define_grid.R
\name{define.grid3}
\alias{define.grid3}
\title{Define AOI native grid}
\usage{
define.grid3(AOI, source = NULL)
}
\arguments{
\item{AOI}{an AOI object}

\item{source}{a datset id (source)  being called}
}
\value{
a list of values including type, lat.call, lon.call, extent, AOI, proj4string
}
\description{
**INTERNAL** This function defines the grid of the AOI in terms of the climate resource
}
\author{
Mike Johnson
}
\keyword{internal}
