% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_define_versions.R
\name{define.versions}
\alias{define.versions}
\title{Define climateR versions}
\usage{
define.versions(
  dates,
  first.date = "1950-01-01",
  sep.date = "2006-01-01",
  scenario,
  future.call,
  historic.call,
  timeRes = "daily"
)
}
\arguments{
\item{dates}{a vector of dates}

\item{first.date}{the first date in a dataset}

\item{sep.date}{the data in which seperates TDS catolouges}

\item{scenario}{the climate scenario}

\item{future.call}{the TDS future catologue}

\item{historic.call}{the TDS historic catologue}

\item{timeRes}{the time resolution of dataset (e.g. monthly or daily)}
}
\value{
data.frame
}
\description{
**INTERNAL** Define if data belongs to separate 
(historic and future) archives based on a defiend seperation dat.
}
