% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_define_dates.R
\name{.numberOfDays}
\alias{.numberOfDays}
\title{Days in Month}
\usage{
.numberOfDays(date)
}
\arguments{
\item{date}{character starting date in the form YYYY-MM-DD.}
}
\value{
the number of days in current month
}
\description{
From a date, return the number of days in that month
}
\keyword{internal}
