% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_define_dates.R
\name{.numberOfMonths}
\alias{.numberOfMonths}
\title{Months between dates}
\usage{
.numberOfMonths(date, baseDate)
}
\arguments{
\item{startDate}{character starting date in the form YYYY-MM-DD.}
}
\value{
the number of days in current month
}
\description{
Find the number of months between dates
}
\keyword{internal}
