% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_define_config.R
\name{.random_model}
\alias{.random_model}
\title{Select N number of GCMs}
\usage{
.random_model(dataset, n = 10)
}
\arguments{
\item{dataset}{the dataset for which random GCMs are being drawn}

\item{n}{the number of model names to return}
}
\value{
a vector of N GCMs
}
\description{
**INTERNAL** Returns a vector of possible GCM models of length N
}
\keyword{internal}
