% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.resource_grid}
\alias{.resource_grid}
\title{Extract grid metadata from NC Pointer}
\usage{
.resource_grid(nc, X_name = NULL, Y_name = NULL, stopIfNotEqualSpaced = TRUE)
}
\arguments{
\item{nc}{"NetCDF" object which points to the NetCDF dataset. Found with RNetCDF::open.nc.}

\item{X_name}{Name of X diminsion. If NULL it is found}

\item{Y_name}{Name of Y diminsion. If NULL it is found}

\item{stopIfNotEqualSpaced}{stop if not equal space grid}
}
\value{
list
}
\description{
Extract grid metadata from NC Pointer
}
\seealso{
Other dap: 
\code{\link{.resource_time}()},
\code{\link{climater_dap}()},
\code{\link{climater_filter}()},
\code{\link{dap_crop}()},
\code{\link{dap_get}()},
\code{\link{dap_meta}()},
\code{\link{dap_summary}()},
\code{\link{dap_to_local}()},
\code{\link{dap_xyzv}()},
\code{\link{dap}()},
\code{\link{extract_sites}()},
\code{\link{get_data}()},
\code{\link{go_get_dap_data}()},
\code{\link{grid_meta}()},
\code{\link{make_ext}()},
\code{\link{make_vect}()},
\code{\link{merge_across_time}()},
\code{\link{parse_date}()},
\code{\link{read_dap_file}()},
\code{\link{read_ftp}()},
\code{\link{time_meta}()},
\code{\link{try_att}()},
\code{\link{var_to_terra}()},
\code{\link{variable_meta}()},
\code{\link{vrt_crop_get}()}
}
\concept{dap}
