% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr.R
\name{.resource_grid_zarr}
\alias{.resource_grid_zarr}
\title{Extract Grid Information from Zarr Resource}
\usage{
.resource_grid_zarr(
  URL,
  X_name = NULL,
  Y_name = NULL,
  stopIfNotEqualSpaced = TRUE
)
}
\arguments{
\item{URL}{Character. The URL of the Zarr file.}

\item{X_name}{Character. Name of the X-coordinate variable. Defaults to NULL.}

\item{Y_name}{Character. Name of the Y-coordinate variable. Defaults to NULL.}

\item{stopIfNotEqualSpaced}{Logical. Whether to stop if grid cells are not equally spaced. Defaults to TRUE.}
}
\value{
A data frame with grid properties.
}
\description{
Retrieves grid dimension metadata and calculates grid properties.
}
\seealso{
Other zarr: 
\code{\link{.resource_time_zarr}()},
\code{\link{go_get_zarr}()},
\code{\link{read_zarr_file}()},
\code{\link{zarr_crop}()},
\code{\link{zarr_get}()},
\code{\link{zarr_to_terra}()},
\code{\link{zarr_xyzv}()}
}
\concept{zarr}
