% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr.R
\name{.resource_time_zarr}
\alias{.resource_time_zarr}
\title{Extract Time Information from Zarr Resource}
\usage{
.resource_time_zarr(URL, T_name = NULL)
}
\arguments{
\item{URL}{Character. The URL of the Zarr file.}

\item{T_name}{Character. Name of the time variable. Defaults to NULL.}
}
\value{
A list with time duration, interval, and count information.
}
\description{
Retrieves time dimension metadata and calculates time intervals.
}
\seealso{
Other zarr: 
\code{\link{.resource_grid_zarr}()},
\code{\link{go_get_zarr}()},
\code{\link{read_zarr_file}()},
\code{\link{zarr_crop}()},
\code{\link{zarr_get}()},
\code{\link{zarr_to_terra}()},
\code{\link{zarr_xyzv}()}
}
\concept{zarr}
