% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-sites.R
\name{extract_sites}
\alias{extract_sites}
\title{Extract Sites}
\usage{
extract_sites(r, pts, id)
}
\arguments{
\item{r}{a raster object}

\item{pts}{point to extract from}

\item{id}{the unique identifier of each point (column name from pts)}
}
\value{
a data.frame with columnes represneting points, and rows time periods
}
\description{
extract timeseries values from a raster stack for a set of points
}
