% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_fast_download.R
\name{fast.download}
\alias{fast.download}
\title{Fast Download for OPenDAP URLs}
\usage{
fast.download(
  urls,
  params,
  names,
  g,
  date.names,
  dataset,
  fun = "r",
  no_data = NA,
  scale_factor = 1
)
}
\arguments{
\item{urls}{a set of OPeNDAP URLS}

\item{params}{a set of parameters}

\item{names}{a set of variables names}

\item{g}{a list of grid data (found with define.grid)}

\item{date.names}{a list of parsed dates}

\item{dataset}{a dataset name (source/id)}

\item{fun}{a function to structure TDS output}

\item{no_data}{a no_date value}

\item{scale_factor}{a scale factors for returned data}
}
\value{
a list object of raster*
}
\description{
Parrallelized download and processing of TDS data
}
