% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_reproject.R
\name{fast_reproject}
\alias{fast_reproject}
\title{Faster Raster Re-projection}
\usage{
fast_reproject(r, target_prj, method = "bilinear")
}
\arguments{
\item{r}{a single or list of rasterLayer or rasterStack objects}

\item{target_prj}{the target projection of the output raster. 
Can be either a proj4string, or an EPSG code}

\item{method}{the method for resampling/reprojecting. Default is 'bilinear'. 
Options can be found [here](https://gdal.org/programs/gdalwarp.html#cmdoption-gdalwarp-r)}
}
\description{
Uses the `sf` GDAL bindings to provide fast raster reprojection. 
Currently we are only supporting the ability to transform the grid reference 
system, not the grid resolution.
}
