% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather.R
\name{get_weather}
\alias{get_weather}
\title{getpastinfo: Minimum, maximum temperatures, average wind speed, and average humidity for a given date of a particular city}
\usage{
get_weather(cities, continent)
}
\arguments{
\item{cities}{string value}

\item{continent}{string value}
}
\value{
map
}
\description{
getpastinfo: Minimum, maximum temperatures, average wind speed, and average humidity for a given date of a particular city
}
\examples{
get_weather("London","Europe")


}
