% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBCSD.R
\name{getBCSDvic}
\alias{getBCSDvic}
\title{Get BCSD Hydrologic (VIC) Outputs for an Area of Interest}
\usage{
getBCSDvic(AOI, param, model = "CCSM4", scenario = "rcp45", startDate,
  endDate = NULL)
}
\arguments{
\item{AOI}{a spatial polygon object (sf or sp)}

\item{param}{a meterological parameter (see `param_meta$bcsd`)}

\item{model}{GMC model name (see `model_meta$bcsd`)}

\item{scenario}{a climate scenario pathway (rcp45 or rcp85)}

\item{startDate}{a start date given as "YYYY-MM-DD"}

\item{endDate}{an end date given as "YYYY-MM-DD"}
}
\value{
if AOI is an areal extent a list of rasterStacks, if AOI is a point then a data.frame of modeled records.
}
\description{
CMIP5 Monthly Hydrology Projections driven with BCSD climate forcings
}
\author{
Mike Johnson
}
