% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCHIRPS.R
\name{getCHIRPS}
\alias{getCHIRPS}
\title{Get CHIRPS Precipitation Data for an Area of Interest}
\usage{
getCHIRPS(AOI, startDate, endDate = NULL)
}
\arguments{
\item{AOI}{a spatial polygon object (sf or sp)}

\item{startDate}{a start date given as "YYYY-MM-DD"}

\item{endDate}{an end date given as "YYYY-MM-DD"}
}
\value{
a list of rasterStacks or dataframe of point values
}
\description{
Climate Hazards Group InfraRed Precipitation with Station data (CHIRPS) is a 30+ year quasi-global rainfall dataset. Spanning 50°S-50°N (and all longitudes), starting in 1981 to near-present,
CHIRPS incorporates 0.05° resolution satellite imagery with in-situ station data to create gridded rainfall time series for trend analysis and seasonal drought monitoring.
}
\author{
Mike Johnson
}
