% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDaymet.R
\name{getDaymet}
\alias{getDaymet}
\title{Get Daymet Climate Data for an Area of Interest}
\usage{
getDaymet(AOI, param, startDate, endDate = NULL)
}
\arguments{
\item{AOI}{a spatial polygon object (sf or sp)}

\item{param}{a meterological parameter (see `param_meta$daymet`)}

\item{startDate}{a start date given as "YYYY-MM-DD"}

\item{endDate}{an end date given as "YYYY-MM-DD"}
}
\value{
if AOI is an areal extent a list of rasterStacks, if AOI is a point then a data.frame of modeled records.
}
\description{
This dataset provides Daymet Version 3 model output data as gridded estimates of daily weather parameters for North America.
Daymet output variables include the following parameters: minimum temperature, maximum temperature, precipitation, shortwave radiation,
vapor pressure, snow water equivalent, and day length. The dataset covers the period from January 1, 1980 to December 31 of the most recent full calendar year. Each subsequent year is processed individually at the close of a calendar year after
allowing adequate time for input weather station data to be of archive quality.  Daymet variables are continuous surfaces provided as individual files, by year, at a 1-km x 1-km spatial resolution and a daily temporal resolution. Data are in a Lambert Conformal
Conic projection for North America and are in a netCDF file format compliant with Climate and Forecast (CF) metadata conventions.
}
\author{
Mike Johnson
}
