% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEDDI.R
\name{getEDDI}
\alias{getEDDI}
\title{Get EDDI Climate Data for an Area of Interest}
\usage{
getEDDI(
  AOI = NULL,
  startDate,
  endDate = NULL,
  timestep = 3,
  timescale = "week"
)
}
\arguments{
\item{AOI}{a spatial polygon object (sf or sp)}

\item{startDate}{a start date given as "YYYY-MM-DD"}

\item{endDate}{an end date given as "YYYY-MM-DD"}

\item{timestep}{The number of time steps (numeric). Must be between 1-12.}

\item{timescale}{Can be 'week' or 'month'}
}
\value{
if AOI is an areal extent a list of rasterStacks, if AOI is a point then a data.frame of modeled records.
}
\description{
The EDDI (Evaporative Demand Drought Index) is an experimental drought monitoring and early warning guidance tool. 
It examines how anomalous the atmospheric evaporative demand (E0; also known as "the thirst of the atmosphere") is for a given location 
across a time period of interest. EDDI is multi-scalar, meaning that this time  period —can vary to capture drying dynamics 
that themselves operate at different timescales. The EDDI for a given date is the aggregated values for a time period before hand. 
The time period is given in either a number of months or number of weeks. For example, a timestep of 3 and a timescale of weeks requests 
the EDDI values for the 3 weeks prior to the requested date.
}
