% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{getGridMET}
\alias{getGridMET}
\title{Get GridMet Climate Data for an Area of Interest}
\usage{
getGridMET(
  AOI,
  varname,
  startDate,
  endDate = NULL,
  verbose = FALSE,
  dryrun = FALSE
)
}
\arguments{
\item{AOI}{an sf point or polygon to extract data for}

\item{varname}{variable name to extract (e.g. tmin)}

\item{startDate}{a start date given as "YYYY-MM-DD" to extract data for}

\item{endDate}{an end date given as "YYYY-MM-DD" to extract data for}

\item{verbose}{Should messages be emited?}

\item{dryrun}{Return summary of data prior to retrieving it}
}
\value{
if AOI is polygon a list of SpatRasters, if AOI is a point then a data.frame of modeled records.
}
\description{
gridMET is a dataset of daily high-spatial resolution (~4-km, 1/24th degree) surface meteorological data covering the contiguous US from 1979-yesterday.
These data are updated daily.
}
