% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{getISRIC_soils}
\alias{getISRIC_soils}
\title{Get World Soil Information gridded weather and climate data for historical (near current) conditions.}
\usage{
getISRIC_soils(AOI = NULL, varname = NULL, verbose = TRUE, ID = NULL)
}
\arguments{
\item{AOI}{an sf of SpatVect point or polygon to extract data for}

\item{varname}{variable name to extract (e.g. tmin)}

\item{verbose}{Should messages be emited?}

\item{ID}{a column of unique identifiers}
}
\value{
if AOI is polygon a list of SpatRasters, if AOI is a point then a data.frame of modeled records.
}
\description{
World Soil Information (International Soil Reference and Information Centre) serves the international community with open access global soil data
}
\seealso{
Other shortcuts: 
\code{\link{get3DEP}()},
\code{\link{getBCCA}()},
\code{\link{getCABCM}()},
\code{\link{getCHIRPS}()},
\code{\link{getDaymet}()},
\code{\link{getGLDAS}()},
\code{\link{getGridMET}()},
\code{\link{getLCMAP}()},
\code{\link{getLOCA}()},
\code{\link{getLOCA_hydro}()},
\code{\link{getLivneh}()},
\code{\link{getLivneh_fluxes}()},
\code{\link{getMACA}()},
\code{\link{getMODIS}()},
\code{\link{getNASADEM}()},
\code{\link{getNLCD}()},
\code{\link{getNLDAS}()},
\code{\link{getPRISM}()},
\code{\link{getTerraClim}()},
\code{\link{getTerraClimNormals}()},
\code{\link{getVIC}()},
\code{\link{getWorldClim}()}
}
\concept{shortcuts}
