% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{getLCMAP}
\alias{getLCMAP}
\title{Get USGS LCMAP}
\usage{
getLCMAP(AOI, year = 2019, type = "primary landcover")
}
\arguments{
\item{AOI}{an sf point or polygon to extract data for}

\item{year}{Land cover product year 1985 - 2019 (default = 2019)}

\item{type}{product type (primary landcover (default), secondary landcover, primary confidence, secondary confidence, cover change, change day, change magniture, model cquality, spectral stability, spectral lastchance)}
}
\value{
if AOI is polygon a list of SpatRasters, if AOI is a point then a data.frame of modeled records.
}
\description{
Land Change Monitoring, Assessment, and Projection
}
\seealso{
Other shortcuts: 
\code{\link{get3DEP}()},
\code{\link{getBCCA}()},
\code{\link{getCHIRPS}()},
\code{\link{getDaymet}()},
\code{\link{getGLDAS}()},
\code{\link{getGridMET}()},
\code{\link{getLOCA_hydro}()},
\code{\link{getLOCA}()},
\code{\link{getLivneh_fluxes}()},
\code{\link{getLivneh}()},
\code{\link{getMACA}()},
\code{\link{getMODIS}()},
\code{\link{getNASADEM}()},
\code{\link{getNLCD}()},
\code{\link{getNLDAS}()},
\code{\link{getPRISM}()},
\code{\link{getTerraClimNormals}()},
\code{\link{getTerraClim}()},
\code{\link{getVIC}()}
}
\concept{shortcuts}
