% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLOCA.R
\name{getLOCA}
\alias{getLOCA}
\title{Get LOCA Climate Data for an Area of Interest}
\usage{
getLOCA(AOI, param, model = "CCSM4", scenario = "rcp45", startDate,
  endDate = NULL)
}
\arguments{
\item{AOI}{a spatial polygon object (sf or sp)}

\item{param}{a meterological parameter (see `param_meta$loca`)}

\item{model}{GMC model name (see `model_meta$loca`)}

\item{scenario}{a climate scenario pathway (rcp45 or rcp85)}

\item{startDate}{a start date given as "YYYY-MM-DD"}

\item{endDate}{an end date given as "YYYY-MM-DD"}
}
\value{
if AOI is an areal extent a list of rasterStacks, if AOI is a point then a data.frame of modeled records.
}
\description{
LOCA is a statistical downscaling technique that uses past history to add improved fine-scale detail to global climate models.
LOCA has been used to downscale 32 global climate models from the CMIP5 archive at a 1/16th degree spatial resolution, covering North America from central Mexico through Southern Canada. The historical period is 1950-2005,
and there are two future scenarios available: RCP 4.5 and RCP 8.5 over the period 2006-2100 (although some models stop in 2099). The variables currently available are daily minimum and maximum temperature, and daily precipitation.
For more information visit: http://loca.ucsd.edu/.
}
\author{
Mike Johnson
}
