% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{getMODIS}
\alias{getMODIS}
\title{Get MODIS data}
\usage{
getMODIS(
  AOI,
  asset = NULL,
  varname = NULL,
  startDate,
  endDate = NULL,
  verbose = FALSE,
  dryrun = FALSE
)
}
\arguments{
\item{AOI}{an sf point or polygon to extract data for}

\item{asset}{The MODIS sensor}

\item{varname}{variable name to extract (e.g. tmin)}

\item{startDate}{a start date given as "YYYY-MM-DD" to extract data for}

\item{endDate}{an end date given as "YYYY-MM-DD" to extract data for}

\item{verbose}{Should messages be emited?}

\item{dryrun}{Return summary of data prior to retrieving it}
}
\value{
if AOI is polygon a list of SpatRasters, if AOI is a point then a data.frame of modeled records.
}
\description{
Get MODIS data
}
\seealso{
Other shortcuts: 
\code{\link{get3DEP}()},
\code{\link{getBCCA}()},
\code{\link{getCHIRPS}()},
\code{\link{getDaymet}()},
\code{\link{getGLDAS}()},
\code{\link{getGridMET}()},
\code{\link{getLCMAP}()},
\code{\link{getLOCA_hydro}()},
\code{\link{getLOCA}()},
\code{\link{getLivneh_fluxes}()},
\code{\link{getLivneh}()},
\code{\link{getMACA}()},
\code{\link{getNASADEM}()},
\code{\link{getNLCD}()},
\code{\link{getNLDAS}()},
\code{\link{getPRISM}()},
\code{\link{getTerraClimNormals}()},
\code{\link{getTerraClim}()},
\code{\link{getVIC}()}
}
\concept{shortcuts}
