% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netrc_utils.R
\name{getNetrcPath}
\alias{getNetrcPath}
\title{Get the default netrc file path}
\usage{
getNetrcPath()
}
\value{
A character vector containing the default netrc file path
}
\description{
Get a default netrc file path
}
\examples{
getNetrcPath()
}
\seealso{
Other netrc: 
\code{\link{getDodsrcPath}()},
\code{\link{writeDodsrc}()},
\code{\link{writeNetrc}()}
}
\concept{netrc}
