% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpastinfo.R
\name{getpastinfo}
\alias{getpastinfo}
\title{getpastinfo: Minimum, maximum temperatures, average wind speed, and average humidity for a given date of a particular city}
\usage{
getpastinfo(city_name = "London", date = "2013/4/27/")
}
\arguments{
\item{city_name}{string value}

\item{date}{string value}
}
\value{
list of minimum, maximum temperatures, average wind speed, and average humidity for a given past date
}
\description{
getpastinfo: Minimum, maximum temperatures, average wind speed, and average humidity for a given date of a particular city
}
\examples{
getpastinfo("London","2013/4/27/")


}
