% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPRISM.R
\name{getPRISM}
\alias{getPRISM}
\title{Get PRISM Climate Data for an Area of Interest}
\usage{
getPRISM(AOI, param, startDate, endDate = NULL)
}
\arguments{
\item{AOI}{a spatial polygon object (sf or sp)}

\item{param}{a meterological parameter (see `param_meta$prism`)}

\item{startDate}{a start date given as "YYYY-MM-DD"}

\item{endDate}{an end date given as "YYYY-MM-DD"}

\item{model}{GMC model name (see `model_meta$prism$name`)}
}
\value{
a list of rasterStacks
}
\description{
The Parameter-elevation Regressions on Independent Slopes Model (PRISM) group gathers climate observations
from a wide range of monitoring networks, applies sophisticated quality control measures, and develops spatial climate datasets to reveal
short- and long-term climate patterns. The resulting datasets incorporate a variety of modeling techniques and are available at multiple spatial/temporal
resolutions, covering the period from 1895 to the present reflects observed patterns of daily near-surface meteorology and simulated changes in GCMs experiments.
}
\author{
Mike Johnson
}
