% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTerraClim.R
\name{getTerraClimNormals}
\alias{getTerraClimNormals}
\title{Get Terra Climate Normals for an Area of Interest}
\usage{
getTerraClimNormals(AOI, param, period = "19812010", month = 1:12)
}
\arguments{
\item{AOI}{a spatial polygon object (sf or sp)}

\item{param}{a meterological parameter (see `param_meta$terraclim`). 
PDSI is not avalaible for any period, and "srad", "vap", "vpd" 
are only avaialble for '19812010' and  '19611990'}

\item{period}{a period of record. Valid options 
include: '19812010' (default), '19611990', '4C' , and '2C'}

\item{month}{numeric. and month or vector of months to access. Default is 1:12}
}
\value{
if AOI is an areal extent a list of rasterStacks, if AOI is a point then a data.frame of modeled records.
}
\description{
These layers from TerraClimate were creating using climatically aided interpolation of monthly anomalies from the CRU Ts4.0 and Japanese 55-year Reanalysis (JRA-55) datasets with WorldClim v2.0 climatologies.
}
