% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopoWX.R
\name{getTopoWX}
\alias{getTopoWX}
\title{Get TopoWX Climate Data for an Area of Interest}
\usage{
getTopoWX(AOI, param, startDate, endDate = NULL, timeRes = "daily")
}
\arguments{
\item{AOI}{a spatial polygon object (sf or sp)}

\item{param}{a meterological parameter (see `param_meta$topox`)}

\item{startDate}{a start date given as "YYYY-MM-DD"}

\item{endDate}{an end date given as "YYYY-MM-DD"}

\item{timeRes}{daily or monthly}
}
\value{
a list of rasterStacks
}
\description{
The TopoWx ("Topography Weather") dataset contains historical 30-arcsec resolution (~800-m) interpolations of daily minimum and maximum topoclimatic air temperature for the conterminous U.S. 
Using both DEM-based variables and MODIS land skin temperature as predictors of air temperature, interpolation procedures include moving window regression kriging and geographically weighted regression. 
To avoid artificial climate trends, all input station data are homogenized using the GHCN/USHCN Pairwise Homogenization Algorithm (https://www.ncdc.noaa.gov/oa/climate/research/ushcn/#phas). 
The following data are available in this archive: 1948-2016 daily and monthly minimum and maximum temperature.
}
\author{
Mike Johnson
}
