% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{getWorldClim}
\alias{getWorldClim}
\title{Get WorlClim gridded weather and climate data for historical (near current) conditions.}
\usage{
getWorldClim(
  AOI = NULL,
  varname = NULL,
  model = "wc2.1_5m",
  month = 1:12,
  ID = NULL,
  verbose = TRUE
)
}
\arguments{
\item{AOI}{an sf of SpatVect point or polygon to extract data for}

\item{varname}{variable name to extract (e.g. tmin)}

\item{model}{GCM model name generating}

\item{month}{numeric. and month or vector of months to access. Default is 1:12}

\item{ID}{a column of unique identifiers}

\item{verbose}{Should messages be emited?}
}
\value{
if AOI is polygon a list of SpatRasters, if AOI is a point then a data.frame of modeled records.
}
\description{
WorldClim is a database of high spatial resolution global weather and climate data. These data can be used for mapping and spatial modeling.
}
\seealso{
Other shortcuts: 
\code{\link{get3DEP}()},
\code{\link{getBCCA}()},
\code{\link{getCHIRPS}()},
\code{\link{getDaymet}()},
\code{\link{getGLDAS}()},
\code{\link{getGridMET}()},
\code{\link{getISRIC_soils}()},
\code{\link{getLCMAP}()},
\code{\link{getLOCA_hydro}()},
\code{\link{getLOCA}()},
\code{\link{getLivneh_fluxes}()},
\code{\link{getLivneh}()},
\code{\link{getMACA}()},
\code{\link{getMODIS}()},
\code{\link{getNASADEM}()},
\code{\link{getNLCD}()},
\code{\link{getNLDAS}()},
\code{\link{getPRISM}()},
\code{\link{getTerraClimNormals}()},
\code{\link{getTerraClim}()},
\code{\link{getVIC}()}
}
\concept{shortcuts}
