% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{go_get_dap_data}
\alias{go_get_dap_data}
\title{Read formated DAP URL as SpatRast}
\usage{
go_get_dap_data(dap)
}
\arguments{
\item{dap}{output from dap_crop}
}
\value{
SpatRast
}
\description{
Read formated DAP URL as SpatRast
}
