% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr.R
\name{go_get_zarr}
\alias{go_get_zarr}
\title{Retrieve data from a Zarr resource}
\usage{
go_get_zarr(zarr, get = TRUE)
}
\arguments{
\item{zarr}{A data frame containing details of the Zarr resource to process. 
Each row should correspond to a single Zarr resource.}

\item{get}{Logical. If `TRUE`, retrieves data; if `FALSE`, returns metadata information.}
}
\value{
If `get = TRUE`, returns the requested data as a matrix. If `get = FALSE`, 
returns a data frame containing metadata information for the resource.
}
\description{
This function retrieves or prepares metadata for data stored in a Zarr format.
}
\examples{
\dontrun{
# Example usage (assuming `zarr` is a properly formatted data frame):
# result <- go_get_zarr(zarr, get = TRUE)
}
}
\seealso{
Other zarr: 
\code{\link{.resource_grid_zarr}()},
\code{\link{.resource_time_zarr}()},
\code{\link{read_zarr_file}()},
\code{\link{zarr_crop}()},
\code{\link{zarr_get}()},
\code{\link{zarr_to_terra}()},
\code{\link{zarr_xyzv}()}
}
\concept{zarr}
