% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data_docs.R
\docType{data}
\name{grid_meta}
\alias{grid_meta}
\title{Grids}
\format{a \code{data.frame} contained source info for
\itemize{
\item 'daymet':    Daily Surface Weather and Climatological Summaries
\item 'gridmet':   Gridded Meteorological Data; Multivariate Adaptive Constructed Analogs
\item 'prism':     Parameter-elevation Regressions on Independent Slopes
\item 'loca':      Localized Constructed Analogs
\item 'terraclim':      Localized Constructed Analogs
\item 'maca':      Localized Constructed Analogs
}}
\usage{
grid_meta
}
\description{
Grid meta data (Min/Max XY, dx/dy, proj) used for constructing subsetting calls to OPeNDAP.
}
\examples{
\dontrun{
 grids = climateR::grid_meta
}

}
\keyword{datasets}
