% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importContinentsEmissionData.R
\name{importContinentsEmissionData}
\alias{importContinentsEmissionData}
\title{Load emission data used for the climate justice tool.}
\usage{
importContinentsEmissionData()
}
\description{
Load emission data used for the climate justice tool.
}
\details{
Loads and prepares files with data containing the continents' emission history.
}
