% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCountriesEmissionData.R
\name{importCountriesEmissionData}
\alias{importCountriesEmissionData}
\title{Load emission data used for the climate justice tool.}
\usage{
importCountriesEmissionData()
}
\description{
Load emission data used for the climate justice tool.
}
\details{
Loads and prepares files with data containing the countries' emission history.
}
