% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importEmissionData.R
\name{importEmissionData}
\alias{importEmissionData}
\title{Load emission data used for the climate justice tool.}
\usage{
importEmissionData()
}
\description{
Load emission data used for the climate justice tool.
}
\details{
Loads and prepares files with data containing the countries' emission history.
}
