% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importIPCCData.R
\name{importIPCCData}
\alias{importIPCCData}
\title{Load IPCC budget data used for the climate justice tool.}
\usage{
importIPCCData()
}
\description{
Load IPCC budget data used for the climate justice tool.
}
\details{
Loads and prepares files with data containing IPCC's emission budgets.
}
