% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadWorldMap.R
\name{loadWorldMap}
\alias{loadWorldMap}
\title{Load World Map}
\usage{
loadWorldMap(quality)
}
\arguments{
\item{quality}{Specifies how detailed the map should be. This determines the processing time for creating a chloropleth graph. Available choices: low, medium, high}
}
\description{
Load World Map
}
\details{
Loads and prepares a geoJSON world map. Returns a SpatialPolygonsDataFrame.
}
