% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonalStats.R
\name{makeGrid}
\alias{makeGrid}
\title{Make Grid}
\usage{
makeGrid(r)
}
\arguments{
\item{r}{a raster object}
}
\value{
a sf layer of polygons
}
\description{
This function creates a simple feature layer from an input raster to be used with buildWeights and zonal
}
\examples{
\dontrun{
AOI = getAOI(state = "CA", county = "all")

r = getMACA(AOI, "prcp", startDate = "2018-10-10")
g = makeGrid(r)
}
}
\author{
Mike Johnson
}
