% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLabelArrows.R
\name{makeLabelArrows}
\alias{makeLabelArrows}
\title{Make label arrows}
\usage{
makeLabelArrows(data, x, xend, y, yend, label)
}
\arguments{
\item{data}{ggplot object}

\item{x}{The place where the arrow start and where the annotation is}

\item{xend}{End of the arrow, where the arrow is pointing to}

\item{y}{The place where the arrow start and where the annotation is}

\item{yend}{End of the arrow, where the arrow is pointing to}

\item{label}{Label text.}
}
\description{
Make label arrows
}
\details{
Adds labels with arrows at the desired place in the plot
}
