% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_across_time}
\alias{merge_across_time}
\title{Merge List of SpatRaster's across time}
\usage{
merge_across_time(data)
}
\arguments{
\item{data}{list of names SpatRasters}
}
\value{
data.frame with (varname, X_name, Y_name, T_name)
}
\description{
Given a list of SpatRasters with possibly shared names, merge across time
}
\seealso{
Other dap: 
\code{\link{.resource_grid}()},
\code{\link{.resource_time}()},
\code{\link{climater_dap}()},
\code{\link{climater_filter}()},
\code{\link{dap}()},
\code{\link{dap_crop}()},
\code{\link{dap_get}()},
\code{\link{dap_meta}()},
\code{\link{dap_summary}()},
\code{\link{dap_to_local}()},
\code{\link{dap_xyzv}()},
\code{\link{extract_sites}()},
\code{\link{get_data}()},
\code{\link{go_get_dap_data}()},
\code{\link{grid_meta}()},
\code{\link{make_ext}()},
\code{\link{make_vect}()},
\code{\link{parse_date}()},
\code{\link{read_dap_file}()},
\code{\link{read_ftp}()},
\code{\link{time_meta}()},
\code{\link{try_att}()},
\code{\link{var_to_terra}()},
\code{\link{variable_meta}()},
\code{\link{vrt_crop_get}()}
}
\concept{dap}
