% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data_docs.R
\docType{data}
\name{model_meta}
\alias{model_meta}
\title{Model Metadata}
\format{An object of class \code{list} of length 3.}
\usage{
model_meta
}
\description{
Metadata for model options
}
\examples{
\dontrun{
 model_meta = climateR::model_meta
}

}
\keyword{datasets}
