% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data_docs.R
\docType{data}
\name{param_meta}
\alias{param_meta}
\title{Model Parameter Metadata}
\format{
An object of class \code{list} of length 12.
}
\usage{
param_meta
}
\description{
Metadata for model parameters. Includes common name, model name, despriction and units
}
\examples{
\dontrun{
 climateR::param_meta
}

}
\keyword{datasets}
