% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap.R
\name{parse_date}
\alias{parse_date}
\title{Parse Dates from duration and interval}
\usage{
parse_date(duration, interval)
}
\arguments{
\item{duration}{time duration}

\item{interval}{time interval}
}
\description{
Parse Dates from duration and interval
}
