% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareBarChartContinents.R
\name{prepareBarChartContinents}
\alias{prepareBarChartContinents}
\title{Bar Chart: Emissions of continents}
\usage{
prepareBarChartContinents(data, theme)
}
\arguments{
\item{data}{Data frame containing the emission data for continents. Needs to have the columns continent, emissions, data_id}
}
\description{
Bar Chart: Emissions of continents
}
\details{
Prepares the bar chart displaying the continents' emissions over time
}
