% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareLineChartContinents.R
\name{prepareLineChartContinents}
\alias{prepareLineChartContinents}
\title{Line Chart: Emissions of continents}
\usage{
prepareLineChartContinents(data, theme)
}
\arguments{
\item{data}{Data frame containing the emission data for continents. Needs to have the columns continent, emissions, data_id}
}
\description{
Line Chart: Emissions of continents
}
\details{
Prepares the line chart displaying the continents' emissions over time
}
