% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareLineChartTemperatures.R
\name{prepareLineChartTemperatures}
\alias{prepareLineChartTemperatures}
\title{Line chart: Development of global temperature}
\usage{
prepareLineChartTemperatures(data, theme)
}
\arguments{
\item{data}{Data frame containing the emission data}
}
\description{
Line chart: Development of global temperature
}
\details{
Prepares the line chart displaying the global temperature anomalies
}
