% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareScatterPlotBudgets.R
\name{prepareScatterPlotBudgets}
\alias{prepareScatterPlotBudgets}
\title{Scatter plot: IPCC Carbon budgets}
\usage{
prepareScatterPlotBudgets(data, theme)
}
\arguments{
\item{data}{Data frame containing the data}
}
\description{
Scatter plot: IPCC Carbon budgets
}
\details{
Prepares the scatter plot showing the carbon budgets left
}
