% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareScatterPlotGDP.R
\name{prepareScatterPlotGDP}
\alias{prepareScatterPlotGDP}
\title{Scatter plot: Emissions and GDP per capita}
\usage{
prepareScatterPlotGDP(data, theme, limits_x, limits_y, limits_scale, colors)
}
\arguments{
\item{data}{Data frame containing the emission data}
}
\description{
Scatter plot: Emissions and GDP per capita
}
\details{
Prepares the scatter plot
}
