% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_dap_file}
\alias{read_dap_file}
\title{Read from a OpenDAP landing page}
\usage{
read_dap_file(URL, varname = NULL, id, varmeta = TRUE)
}
\arguments{
\item{URL}{URL to OpenDap resource}

\item{id}{character. Uniquely named dataset identifier}

\item{varmeta}{should variable metadata be appended?}
}
\value{
data.frame
}
\description{
Reads an OpenDap resources and returns metadata
}
