% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap.R
\name{read_ftp}
\alias{read_ftp}
\title{Read from FTP}
\usage{
read_ftp(URL, cat, lyrs = 1, AOI, ext = NULL, crs = NULL, dates = NULL)
}
\arguments{
\item{cat}{catalog element}

\item{lyrs}{lyrs to extract}

\item{AOI}{Area of Interest}

\item{ext}{extent of source (if needed)}

\item{crs}{crs of source (if needed)}

\item{dates}{dates of data}

\item{url}{Unique Resource Identifier (http or local)}
}
\value{
SpatRaster
}
\description{
Read from FTP
}
