% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_catalog.R
\name{read_live_catalog}
\alias{read_live_catalog}
\title{Read Live Catalog from Github release}
\usage{
read_live_catalog(
  url = paste0("https://github.com/mikejohnson51/climateR-catalogs",
    "/releases/latest/download/", "catalog.parquet")
)
}
\arguments{
\item{url}{URL to read}
}
\value{
data.frame
}
\description{
Every month, our data catalog is refreshed. This function reads 
the most current catalog from the Github release.
}
\seealso{
Other catalog: 
\code{\link{catalog}}
}
\concept{catalog}
