% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr.R
\name{read_zarr_file}
\alias{read_zarr_file}
\title{Read Zarr File}
\usage{
read_zarr_file(URL, varname = NULL, id, varmeta = TRUE)
}
\arguments{
\item{URL}{Character. The URL of the Zarr file.}

\item{varname}{Character. Variable name to extract. Defaults to NULL.}

\item{id}{Character. An identifier for the dataset.}

\item{varmeta}{Logical. Whether to include variable metadata. Defaults to TRUE.}
}
\value{
A data frame with merged metadata and variable information.
}
\description{
Reads a Zarr file from a specified URL and extracts metadata and variable information.
}
\seealso{
Other zarr: 
\code{\link{.resource_grid_zarr}()},
\code{\link{.resource_time_zarr}()},
\code{\link{go_get_zarr}()},
\code{\link{zarr_crop}()},
\code{\link{zarr_get}()},
\code{\link{zarr_to_terra}()},
\code{\link{zarr_xyzv}()}
}
\concept{zarr}
