% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suntime.R
\name{suntime}
\alias{suntime}
\title{suntime: Sunrise and sunset time of a particular city}
\usage{
suntime(city)
}
\arguments{
\item{city}{string value}
}
\value{
list of sunrise and sunset time in the given city
}
\description{
suntime: Sunrise and sunset time of a particular city
}
\examples{
get_suntime("London")
}
