% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try_att}
\alias{try_att}
\title{TryCatch around RNetCDF::att.get.nc()}
\usage{
try_att(nc, variable, attribute)
}
\arguments{
\item{nc}{"NetCDF" object which points to the NetCDF dataset. Found with RNetCDF::open.nc.}

\item{variable}{ID or name of the variable from which the attribute will be read, or "NC_GLOBAL" for a global attribute.}

\item{attribute}{Attribute name or ID.}
}
\value{
Vector with a data type that depends on the NetCDF variable. For NetCDF variables of type NC_CHAR, the R type is either character or raw, as specified by argument rawchar. For NC_STRING, the R type is character. Numeric variables are read as double precision by default, but the smallest R type that exactly represents each external type is used if fitnum is TRUE.
}
\description{
TryCatch around RNetCDF::att.get.nc()
}
\seealso{
Other dap: 
\code{\link{.resource_grid}()},
\code{\link{.resource_time}()},
\code{\link{climater_dap}()},
\code{\link{climater_filter}()},
\code{\link{dap_crop}()},
\code{\link{dap_get}()},
\code{\link{dap_meta}()},
\code{\link{dap_summary}()},
\code{\link{dap_to_local}()},
\code{\link{dap_xyzv}()},
\code{\link{dap}()},
\code{\link{extract_sites}()},
\code{\link{get_data}()},
\code{\link{go_get_dap_data}()},
\code{\link{grid_meta}()},
\code{\link{make_ext}()},
\code{\link{make_vect}()},
\code{\link{merge_across_time}()},
\code{\link{parse_date}()},
\code{\link{read_dap_file}()},
\code{\link{read_ftp}()},
\code{\link{time_meta}()},
\code{\link{var_to_terra}()},
\code{\link{variable_meta}()},
\code{\link{vrt_crop_get}()}
}
\concept{dap}
