% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_to_terra}
\alias{var_to_terra}
\title{Variable Array to SpatRast}
\usage{
var_to_terra(var, dap)
}
\arguments{
\item{var}{numeric array}

\item{dap}{dap description}
}
\value{
SpatRast
}
\description{
Variable Array to SpatRast
}
