% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_tools.R
\name{variable_meta}
\alias{variable_meta}
\title{Find DAP variable metadata}
\usage{
variable_meta(raw, verbose = TRUE)
}
\arguments{
\item{raw}{data.frame}

\item{verbose}{emit messages}
}
\value{
data.frame
}
\description{
Find DAP variable metadata
}
