% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap.R
\name{vrt_crop_get}
\alias{vrt_crop_get}
\title{VRT Crop}
\usage{
vrt_crop_get(
  URL = NULL,
  catalog = NULL,
  AOI = NULL,
  grid = NULL,
  varname = NULL,
  start = NULL,
  end = NULL,
  toptobottom = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{URL}{local file path or URL}

\item{catalog}{subset of open.dap catalog}

\item{AOI}{an sf point or polygon to extract data for}

\item{grid}{a list containing an extent (), and crs}

\item{varname}{variable name to extract (e.g. tmin)}

\item{start}{for non "dated" items, start can be called by index}

\item{end}{for non "dated" items, end can be called by index}

\item{toptobottom}{should data be inverse?}

\item{verbose}{Should dap_summary be printed?}
}
\value{
SpatRaster
}
\description{
VRT Crop
}
