% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netrc_utils.R
\name{writeDodsrc}
\alias{writeDodsrc}
\title{Write dodsrc file}
\usage{
writeDodsrc(
  netrcFile = getNetrcPath(),
  dodsrcFile = getDodsrcPath(),
  overwrite = FALSE
)
}
\arguments{
\item{netrcFile}{A character. A path to where the netrc file should be.}

\item{dodsrcFile}{the machine you are logging into
By default will go to your home directory, which is advised}

\item{overwrite}{A logical. overwrite the existing dodsrc file?}
}
\value{
A character vector containing the netrc file path
}
\description{
Write a dodsrc file that is valid for a netrc file
}
\examples{
\dontrun{
writeNetrc(
  login = "XXX@email.com",
  password = "yourSecretPassword"
)
}
}
\seealso{
\code{\link{checkNetrc}}

Other netrc: 
\code{\link{getDodsrcPath}()},
\code{\link{getNetrcPath}()},
\code{\link{writeNetrc}()}
}
\concept{netrc}
