% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netrc_utils.R
\name{writeNetrc}
\alias{writeNetrc}
\title{Write netrc file}
\usage{
writeNetrc(
  login,
  password,
  machine = "urs.earthdata.nasa.gov",
  netrcFile = getNetrcPath(),
  overwrite = FALSE
)
}
\arguments{
\item{login}{A character. Email address used for logging in on earthdata}

\item{password}{A character. Password associated with the login.}

\item{machine}{the machine you are logging into}

\item{netrcFile}{A character. A path to where the netrc file should be written.
By default will go to your home directory, which is advised}

\item{overwrite}{A logical. overwrite the existing netrc file?}
}
\value{
A character vector containing the netrc file path
}
\description{
Write a netrc file that is valid for accessing urs.earthdata.nasa.gov
}
\details{
The database is accessed with the user's credentials.
A netrc file storing login and password information is required.
See \href{https://urs.earthdata.nasa.gov/}{here}. Once set up you must do the following
(1) Login to EarthData (2) Go to Applications > Authorized Apps (3) If NASA GESDISC DATA ARCHIVE is not in the Approved Applications list, select APPROVE MORE APPLICATIONS
(4) Find NASA GESDISC DATA ARCHIVE and click AUTHORIZE
for instruction on how to register and set DataSpace credential.
}
\examples{
\dontrun{
writeNetrc(
  login = "XXX@email.com",
  password = "yourSecretPassword"
)
}
}
\seealso{
\code{\link{checkNetrc}}

Other netrc: 
\code{\link{checkDodsrc}()},
\code{\link{checkNetrc}()},
\code{\link{getDodsrcPath}()},
\code{\link{getNetrcPath}()},
\code{\link{writeDodsrc}()}
}
\concept{netrc}
