% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr.R
\name{zarr_crop}
\alias{zarr_crop}
\title{Crop Zarr Data to a Spatial and Temporal Subset}
\usage{
zarr_crop(
  URL = NULL,
  catalog = NULL,
  AOI = NULL,
  startDate = NULL,
  endDate = NULL,
  start = NULL,
  end = NULL,
  varname = NULL,
  verbose = TRUE
)
}
\arguments{
\item{URL}{Character. The URL of the Zarr file. Defaults to NULL.}

\item{catalog}{Data frame. Metadata catalog for the Zarr file. Defaults to NULL.}

\item{AOI}{Spatial object. Area of interest for cropping. Defaults to NULL.}

\item{startDate}{Character. Start date for cropping. Defaults to NULL.}

\item{endDate}{Character. End date for cropping. Defaults to NULL.}

\item{start}{Numeric. Start index for cropping. Defaults to NULL.}

\item{end}{Numeric. End index for cropping. Defaults to NULL.}

\item{varname}{Character. Variable name to crop. Defaults to NULL.}

\item{verbose}{Logical. Whether to print verbose output. Defaults to TRUE.}
}
\value{
A cropped dataset matching the specified criteria.
}
\description{
Crops data in a Zarr file based on spatial (AOI) and temporal (start/end) filters.
}
\seealso{
Other zarr: 
\code{\link{.resource_grid_zarr}()},
\code{\link{.resource_time_zarr}()},
\code{\link{go_get_zarr}()},
\code{\link{read_zarr_file}()},
\code{\link{zarr_get}()},
\code{\link{zarr_to_terra}()},
\code{\link{zarr_xyzv}()}
}
\concept{zarr}
