% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr.R
\name{zarr_to_terra}
\alias{zarr_to_terra}
\title{Convert data to terra SpatRaster or data frame format}
\usage{
zarr_to_terra(var, zarr)
}
\arguments{
\item{var}{A variable containing extracted Zarr data.}

\item{zarr}{A data frame containing details of the Zarr resource, including metadata.}
}
\value{
Returns a terra SpatRaster object or data frame containing the processed data.
}
\description{
Converts extracted Zarr data into terra SpatRaster objects or data frames based on 
spatial and temporal dimensions.
}
\examples{
\dontrun{
# Example usage (assuming `var` and `zarr` are properly formatted):
# result <- zarr_to_terra(var, zarr)
}
}
\seealso{
Other zarr: 
\code{\link{.resource_grid_zarr}()},
\code{\link{.resource_time_zarr}()},
\code{\link{go_get_zarr}()},
\code{\link{read_zarr_file}()},
\code{\link{zarr_crop}()},
\code{\link{zarr_get}()},
\code{\link{zarr_xyzv}()}
}
\concept{zarr}
