% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr.R
\name{zarr_xyzv}
\alias{zarr_xyzv}
\title{Extract Variable Information from Zarr Object}
\usage{
zarr_xyzv(obj, varname = NULL, varmeta = FALSE)
}
\arguments{
\item{obj}{Zarr object or file path.}

\item{varname}{Character. Specific variable name to extract. Defaults to NULL.}

\item{varmeta}{Logical. Whether to include variable metadata. Defaults to FALSE.}
}
\value{
A data frame containing variable metadata and coordinate information.
}
\description{
Extracts variable and coordinate information from a Zarr object.
}
\seealso{
Other zarr: 
\code{\link{.resource_grid_zarr}()},
\code{\link{.resource_time_zarr}()},
\code{\link{go_get_zarr}()},
\code{\link{read_zarr_file}()},
\code{\link{zarr_crop}()},
\code{\link{zarr_get}()},
\code{\link{zarr_to_terra}()}
}
\concept{zarr}
