% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonalStats.R
\name{zonal}
\alias{zonal}
\title{Compute Zonal Averaging}
\usage{
zonal(grid, poly, w)
}
\arguments{
\item{grid}{a sf data.frame}

\item{poly}{a sf data.frame}

\item{w}{a weighted intersection data.frame}
}
\value{
a sf layer of polygons
}
\description{
Returns a simple features layer with average values by polygon featrue
}
\examples{
\dontrun{
AOI = getAOI(state = "CA", county = "all")

r = getMACA(AOI, "prcp", startDate = "2018-10-10")
g = makeGrid(r)
w = buildWeights(g, AOI, "countyfp")
}
}
\author{
Mike Johnson
}
