% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\docType{data}
\name{catalog}
\alias{catalog}
\title{ClimateR Catalog}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 110983 rows and 29 columns.
}
\usage{
catalog
}
\description{
ClimateR Catalog
}
\seealso{
Other catalog: 
\code{\link{read_live_catalog}()}
}
\concept{catalog}
\keyword{datasets}
