% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_climateR.R
\docType{package}
\name{climateR}
\alias{climateR}
\title{climateR}
\description{
\code{climateR} package
}
\details{
Download gridded climate data

See the README on [github](https://github.com/mikejohnson51/climateR)
}
