% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{climater_dap}
\alias{climater_dap}
\title{ClimateR dry}
\usage{
climater_dap(id, args, verbose, dryrun, print.arg = FALSE)
}
\arguments{
\item{id}{The resource name, agency, or catalog identifier}

\item{args}{The parent function arguments}

\item{verbose}{Should messages be emited?}

\item{dryrun}{Return summary of data prior to retriving it}
}
\value{
data.frame
}
\description{
ClimateR dry
}
