% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climater_filter.R
\name{climater_filter}
\alias{climater_filter}
\title{ClimateR Catalog Filter}
\usage{
climater_filter(
  id = NULL,
  asset = NULL,
  AOI = NULL,
  startDate = NULL,
  endDate = NULL,
  varname = NULL,
  model = NULL,
  scenario = NULL,
  ensemble = NULL
)
}
\arguments{
\item{id}{The resource, agency, or catalog identifier}

\item{asset}{The subdataset or asset in a given resource}

\item{AOI}{an sf point or polygon to extract data for}

\item{startDate}{a start date given as "YYYY-MM-DD" to extract data for}

\item{endDate}{an end date given as "YYYY-MM-DD" to extract data for}

\item{varname}{variable name to extract (e.g. tmin)}

\item{model}{GCM model name generating}

\item{scenario}{A climate or modeling scenario used}

\item{ensemble}{The model ensemble member used to generate data}
}
\value{
data.frame
}
\description{
Filter the climateR catalog based on a set of constraints
}
\seealso{
Other dap: 
\code{\link{.resource_grid}()},
\code{\link{.resource_time}()},
\code{\link{climater_dap}()},
\code{\link{dap_crop}()},
\code{\link{dap_get}()},
\code{\link{dap_meta}()},
\code{\link{dap_summary}()},
\code{\link{dap_to_local}()},
\code{\link{dap_xyzv}()},
\code{\link{dap}()},
\code{\link{extract_sites}()},
\code{\link{get_data}()},
\code{\link{go_get_dap_data}()},
\code{\link{grid_meta}()},
\code{\link{make_ext}()},
\code{\link{make_vect}()},
\code{\link{merge_across_time}()},
\code{\link{parse_date}()},
\code{\link{read_dap_file}()},
\code{\link{read_ftp}()},
\code{\link{time_meta}()},
\code{\link{try_att}()},
\code{\link{var_to_terra}()},
\code{\link{variable_meta}()},
\code{\link{vrt_crop_get}()}
}
\concept{dap}
