% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createChartScales.R
\name{createChartScales}
\alias{createChartScales}
\title{Create Chart scales}
\usage{
createChartScales(data)
}
\arguments{
\item{data}{Atomic numeric vector}
}
\description{
Create Chart scales
}
\details{
Creates numbers used for a color scale in charts. If the vector contains negative values, seven colors are used, center is zero. Otherwise, different colors and center is not zero.
}
