% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap.R
\name{dap}
\alias{dap}
\title{Get Data (Data Access Protocol)}
\usage{
dap(
  URL = NULL,
  catalog = NULL,
  AOI = NULL,
  startDate = NULL,
  endDate = NULL,
  varname = NULL,
  grid = NULL,
  start = NULL,
  end = NULL,
  toptobottom = FALSE,
  ID = NULL,
  verbose = TRUE
)
}
\arguments{
\item{URL}{local file path or URL}

\item{catalog}{subset of open.dap catalog}

\item{AOI}{an sf of SpatVect point or polygon to extract data for}

\item{startDate}{a start date given as "YYYY-MM-DD" to extract data for}

\item{endDate}{an end date given as "YYYY-MM-DD" to extract data for}

\item{varname}{variable name to extract (e.g. tmin)}

\item{grid}{a list containing an extent (), and crs}

\item{start}{for non "dated" items, start can be called by index}

\item{end}{for non "dated" items, end can be called by index}

\item{toptobottom}{should data be inverse?}

\item{ID}{a column of unique identifiers}

\item{verbose}{Should dap_summary be printed?}
}
\value{
data.frame
}
\description{
this function provides a consistent data access protocol (dap) to a wide
range of local and remote resources including VRT, TDS, NetCDF

Define and get data from a DAP resource
}
\details{
Wraps dap_get and dap_crop into one.
If AOI is NULL no spatial crop is executed. If startDate AND endDate are NULL, no temporal crop is executed. If just endDate is NULL it defaults to the startDate.
}
\seealso{
Other dap: 
\code{\link{.resource_grid}()},
\code{\link{.resource_time}()},
\code{\link{climater_dap}()},
\code{\link{climater_filter}()},
\code{\link{dap_crop}()},
\code{\link{dap_get}()},
\code{\link{dap_meta}()},
\code{\link{dap_summary}()},
\code{\link{dap_to_local}()},
\code{\link{dap_xyzv}()},
\code{\link{extract_sites}()},
\code{\link{get_data}()},
\code{\link{go_get_dap_data}()},
\code{\link{grid_meta}()},
\code{\link{make_ext}()},
\code{\link{make_vect}()},
\code{\link{merge_across_time}()},
\code{\link{parse_date}()},
\code{\link{read_dap_file}()},
\code{\link{read_ftp}()},
\code{\link{time_meta}()},
\code{\link{try_att}()},
\code{\link{var_to_terra}()},
\code{\link{variable_meta}()},
\code{\link{vrt_crop_get}()}
}
\concept{dap}
