% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap.R
\name{dap_crop}
\alias{dap_crop}
\title{Crop DAP file}
\usage{
dap_crop(
  URL = NULL,
  catalog = NULL,
  AOI = NULL,
  startDate = NULL,
  endDate = NULL,
  varname = NULL,
  verbose = TRUE
)
}
\arguments{
\item{URL}{local file path or URL}

\item{catalog}{subset of open.dap catalog}

\item{AOI}{an sf point or polygon to extract data for}

\item{startDate}{a start date given as "YYYY-MM-DD" to extract data for}

\item{endDate}{an end date given as "YYYY-MM-DD" to extract data for}

\item{varname}{variable name to extract (e.g. tmin)}

\item{verbose}{Should dap_summary be printed?}
}
\value{
data.frame
}
\description{
Crop DAP file
}
