% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap.R
\name{dap_get}
\alias{dap_get}
\title{Get DAP resource data}
\usage{
dap_get(dap, varname = NULL)
}
\arguments{
\item{dap}{data.frame from catalog or dap_crop}

\item{varname}{name of variable to extract. If NULL, then get all}
}
\value{
SpatRaster
}
\description{
Get DAP resource data
}
\seealso{
Other dap: 
\code{\link{.resource_grid}()},
\code{\link{.resource_time}()},
\code{\link{climater_dap}()},
\code{\link{climater_filter}()},
\code{\link{dap_crop}()},
\code{\link{dap_meta}()},
\code{\link{dap_summary}()},
\code{\link{dap_to_local}()},
\code{\link{dap_xyzv}()},
\code{\link{dap}()},
\code{\link{extract_sites}()},
\code{\link{get_data}()},
\code{\link{go_get_dap_data}()},
\code{\link{grid_meta}()},
\code{\link{make_ext}()},
\code{\link{make_vect}()},
\code{\link{merge_across_time}()},
\code{\link{parse_date}()},
\code{\link{read_dap_file}()},
\code{\link{read_ftp}()},
\code{\link{time_meta}()},
\code{\link{try_att}()},
\code{\link{var_to_terra}()},
\code{\link{variable_meta}()},
\code{\link{vrt_crop_get}()}
}
\concept{dap}
