% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap.R
\name{dap_summary}
\alias{dap_summary}
\title{Print Summary Information About a OpenDAP Resource}
\usage{
dap_summary(dap = NULL, url = NULL)
}
\arguments{
\item{dap}{data.frame from catalog or dap_crop}

\item{url}{Unique Resource Identifier (http or local)}
}
\description{
Print summary information about a DAP summary
}
