% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_define_config.R
\name{define.config}
\alias{define.config}
\title{Define climateR configuration}
\usage{
define.config(dataset = NULL, model, ensemble = NA)
}
\arguments{
\item{dataset}{the dataset for which a configuration is needed}

\item{model}{the GCM name(s) of interest}

\item{ensemble}{the number of model names to return}
}
\value{
a vector of N GCMs
}
\description{
**INTERNAL** Define the model configuration to call. 
Ensures that the specified GCM is provided for called dataset
}
\keyword{internal}
