% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.resource_time}
\alias{.resource_time}
\title{Extract time metadata from NC Pointer}
\usage{
.resource_time(URL, T_name = NULL)
}
\arguments{
\item{URL}{location of data to process}

\item{T_name}{Name of T dimension. If NULL it is found}
}
\value{
list
}
\description{
Extract time metadata from NC Pointer
}
\seealso{
Other dap: 
\code{\link{.resource_grid}()},
\code{\link{climater_dap}()},
\code{\link{climater_filter}()},
\code{\link{dap_crop}()},
\code{\link{dap_get}()},
\code{\link{dap_meta}()},
\code{\link{dap_summary}()},
\code{\link{dap_to_local}()},
\code{\link{dap_xyzv}()},
\code{\link{dap}()},
\code{\link{extract_sites}()},
\code{\link{get_data}()},
\code{\link{go_get_dap_data}()},
\code{\link{grid_meta}()},
\code{\link{make_ext}()},
\code{\link{make_vect}()},
\code{\link{merge_across_time}()},
\code{\link{parse_date}()},
\code{\link{read_dap_file}()},
\code{\link{read_ftp}()},
\code{\link{time_meta}()},
\code{\link{try_att}()},
\code{\link{var_to_terra}()},
\code{\link{variable_meta}()},
\code{\link{vrt_crop_get}()}
}
\concept{dap}
