% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{get3DEP}
\alias{get3DEP}
\title{Get USGS 3DEP DEMs}
\usage{
get3DEP(AOI, resolution = "30m")
}
\arguments{
\item{AOI}{an sf point or polygon to extract data for}

\item{resolution}{DEM resolution (10m or 30m (default))}
}
\value{
if AOI is polygon a list of SpatRasters, if AOI is a point then a data.frame of modeled records.
}
\description{
Get USGS 3DEP DEMs
}
\seealso{
Other shortcuts: 
\code{\link{getBCCA}()},
\code{\link{getCHIRPS}()},
\code{\link{getDaymet}()},
\code{\link{getGLDAS}()},
\code{\link{getGridMET}()},
\code{\link{getLCMAP}()},
\code{\link{getLOCA_hydro}()},
\code{\link{getLOCA}()},
\code{\link{getLivneh_fluxes}()},
\code{\link{getLivneh}()},
\code{\link{getMACA}()},
\code{\link{getMODIS}()},
\code{\link{getNASADEM}()},
\code{\link{getNLCD}()},
\code{\link{getNLDAS}()},
\code{\link{getPRISM}()},
\code{\link{getTerraClimNormals}()},
\code{\link{getTerraClim}()},
\code{\link{getVIC}()}
}
\concept{shortcuts}
