% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCABCM.R
\name{getCABCM}
\alias{getCABCM}
\title{Get CABCM Climate Data for an Area of Interest}
\usage{
getCABCM(
  AOI,
  param,
  model = "CCSM4",
  scenario = "rcp85",
  startDate,
  endDate = NULL
)
}
\arguments{
\item{AOI}{a spatial polygon object (sf or sp)}

\item{param}{a meterological parameter (see `param_meta$cabcm``)}

\item{model}{GMC model name (see `model_meta$cabcm`)}

\item{scenario}{a climate scenario pathway (see `model_meta$cabcm`)}

\item{startDate}{a start date given as "YYYY-MM-DD"}

\item{endDate}{an end date given as "YYYY-MM-DD"}
}
\value{
if AOI is an areal extent a list of rasterStacks, if AOI is a point then a data.frame of modeled records.
}
\description{
The Basin Characterization Model (BCM) dataset provides historical and projected climate and hydrology data at a 270 meter resolution,
which is relevant for watershed-scale evaluation and planning.
These data have formed the basis for multiple research projects and vulnerability assessments applying
climate change projections to conservation decision-making, providing a common base-layer and set of
assumptions across these projects.
}
\author{
Mike Johnson
}
