% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netrc_utils.R
\name{getDodsrcPath}
\alias{getDodsrcPath}
\title{Get a default dodsrc file path}
\usage{
getDodsrcPath()
}
\value{
A character vector containing the default netrc file path
}
\description{
Get a default dodsrc file path
}
\examples{
getDodsrcPath()
}
\seealso{
Other netrc: 
\code{\link{getNetrcPath}()},
\code{\link{writeDodsrc}()},
\code{\link{writeNetrc}()}
}
\concept{netrc}
