% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGridMET.R
\name{getGridMET}
\alias{getGridMET}
\title{Get GridMet Climate Data for an Area of Interest}
\usage{
getGridMET(AOI, param, startDate, endDate = NULL)
}
\arguments{
\item{AOI}{a spatial polygon object (sf or sp)}

\item{param}{a meterological parameter (see `param_meta$gridmet`)}

\item{startDate}{a start date given as "YYYY-MM-DD"}

\item{endDate}{an end date given as "YYYY-MM-DD"}
}
\value{
if AOI is an areal extent a list of rasterStacks, if AOI is a point then a data.frame of modeled records.
}
\description{
gridMET is a dataset of daily high-spatial resolution (~4-km, 1/24th degree) surface meteorological data covering the contiguous US from 1979-yesterday.
These data are updated daily.
}
\author{
Mike Johnson
}
