% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{getMACA}
\alias{getMACA}
\title{Get MACA Climate Data for an Area of Interest}
\usage{
getMACA(
  AOI,
  varname,
  timeRes = "day",
  model = "CCSM4",
  scenario = "rcp45",
  startDate,
  endDate = NULL,
  verbose = FALSE,
  dryrun = FALSE
)
}
\arguments{
\item{AOI}{an sf point or polygon to extract data for}

\item{varname}{variable name to extract (e.g. tmin)}

\item{timeRes}{daily or monthly}

\item{model}{GCM model name generating}

\item{scenario}{A climate or modeling scenario used}

\item{startDate}{a start date given as "YYYY-MM-DD" to extract data for}

\item{endDate}{an end date given as "YYYY-MM-DD" to extract data for}

\item{verbose}{Should messages be emited?}

\item{dryrun}{Return summary of data prior to retrieving it}
}
\value{
if AOI is polygon a list of SpatRasters, if AOI is a point then a data.frame of modeled records.
}
\description{
Multivariate Adaptive Constructed Analogs (MACA) is a statistical method for downscaling Global Climate Models
(GCMs) from their native coarse resolution to a higher spatial resolution that captures reflects observed patterns of daily near-surface meteorology and simulated changes in GCMs experiments.
}
\seealso{
Other shortcuts: 
\code{\link{get3DEP}()},
\code{\link{getBCCA}()},
\code{\link{getCHIRPS}()},
\code{\link{getDaymet}()},
\code{\link{getGLDAS}()},
\code{\link{getGridMET}()},
\code{\link{getLCMAP}()},
\code{\link{getLOCA_hydro}()},
\code{\link{getLOCA}()},
\code{\link{getLivneh_fluxes}()},
\code{\link{getLivneh}()},
\code{\link{getMODIS}()},
\code{\link{getNASADEM}()},
\code{\link{getNLCD}()},
\code{\link{getNLDAS}()},
\code{\link{getPRISM}()},
\code{\link{getTerraClimNormals}()},
\code{\link{getTerraClim}()},
\code{\link{getVIC}()}
}
\concept{shortcuts}
