% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMACA.R
\name{getMACA}
\alias{getMACA}
\title{Get MACA Climate Data for an Area of Interest}
\usage{
getMACA(AOI, param, model = "CCSM4", scenario = "rcp45", startDate,
  endDate = NULL, timeRes = "daily")
}
\arguments{
\item{AOI}{a spatial polygon object (sf or sp)}

\item{param}{a meterological parameter (see `param_meta$maca`)}

\item{model}{GMC model name (see `model_meta$maca`)}

\item{scenario}{a climate scenario pathway (rcp45 or rcp85)}

\item{startDate}{a start date given as "YYYY-MM-DD"}

\item{endDate}{an end date given as "YYYY-MM-DD"}

\item{timeRes}{daily or monthly}
}
\value{
a list of rasterStacks
}
\description{
Multivariate Adaptive Constructed Analogs (MACA) is a statistical method for downscaling Global Climate Models
(GCMs) from their native coarse resolution to a higher spatial resolution that catures reflects observed patterns of daily near-surface meteorology and simulated changes in GCMs experiments.
}
\author{
Mike Johnson
}
