% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{getNLCD}
\alias{getNLCD}
\title{Get USGS National Land Cover Dataset}
\usage{
getNLCD(AOI, year = 2019, type = "land cover", ID = NULL)
}
\arguments{
\item{AOI}{an sf of SpatVect point or polygon to extract data for}

\item{year}{Landcover product year (2001, 2011,2016,2019)}

\item{type}{product type}

\item{ID}{a column of unique identifiers}
}
\value{
if AOI is polygon a list of SpatRasters, if AOI is a point then a data.frame of modeled records.
}
\description{
Get USGS National Land Cover Dataset
}
\seealso{
Other shortcuts: 
\code{\link{get3DEP}()},
\code{\link{getBCCA}()},
\code{\link{getCHIRPS}()},
\code{\link{getDaymet}()},
\code{\link{getGLDAS}()},
\code{\link{getGridMET}()},
\code{\link{getISRIC_soils}()},
\code{\link{getLCMAP}()},
\code{\link{getLOCA_hydro}()},
\code{\link{getLOCA}()},
\code{\link{getLivneh_fluxes}()},
\code{\link{getLivneh}()},
\code{\link{getMACA}()},
\code{\link{getMODIS}()},
\code{\link{getNASADEM}()},
\code{\link{getNLDAS}()},
\code{\link{getPRISM}()},
\code{\link{getTerraClimNormals}()},
\code{\link{getTerraClim}()},
\code{\link{getVIC}()},
\code{\link{getWorldClim}()}
}
\concept{shortcuts}
