% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{getTerraClimNormals}
\alias{getTerraClimNormals}
\title{Get TerraClimate Normals for an Area of Interest}
\usage{
getTerraClimNormals(
  AOI,
  varname,
  scenario = "19812010",
  month = 1:12,
  verbose = FALSE,
  ID = NULL,
  dryrun = FALSE
)
}
\arguments{
\item{AOI}{an sf of SpatVect point or polygon to extract data for}

\item{varname}{variable name to extract (e.g. tmin)}

\item{scenario}{A climate or modeling scenario}

\item{month}{numeric. and month or vector of months to access. Default is 1:12}

\item{verbose}{Should messages be emited?}

\item{ID}{a column of unique identifiers}

\item{dryrun}{Return summary of data prior to retrieving it}
}
\value{
if AOI is polygon a list of SpatRasters, if AOI is a point then a data.frame of modeled records.
}
\description{
These layers from TerraClimate were creating using climatically aided interpolation of monthly anomalies from the CRU Ts4.0 and Japanese 55-year Reanalysis (JRA-55) datasets with WorldClim v2.0 climatologies.
}
\seealso{
Other shortcuts: 
\code{\link{get3DEP}()},
\code{\link{getBCCA}()},
\code{\link{getCABCM}()},
\code{\link{getCHIRPS}()},
\code{\link{getDaymet}()},
\code{\link{getGLDAS}()},
\code{\link{getGridMET}()},
\code{\link{getISRIC_soils}()},
\code{\link{getLCMAP}()},
\code{\link{getLOCA_hydro}()},
\code{\link{getLOCA}()},
\code{\link{getLivneh_fluxes}()},
\code{\link{getLivneh}()},
\code{\link{getMACA}()},
\code{\link{getMODIS}()},
\code{\link{getNASADEM}()},
\code{\link{getNLCD}()},
\code{\link{getNLDAS}()},
\code{\link{getPRISM}()},
\code{\link{getTerraClim}()},
\code{\link{getVIC}()},
\code{\link{getWorldClim}()}
}
\concept{shortcuts}
