% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap.R, R/utils.R
\name{make_ext}
\alias{make_ext}
\title{Convert catalog entry to extent}
\usage{
make_ext(cat)

make_ext(cat)
}
\arguments{
\item{cat}{catalog entry (data.frame with an {Xn, X1, Yn, Y1, proj})}
}
\value{
SpatExtent

SpatExtent
}
\description{
Convert catalog entry to extent

Convert catalog entry to extent
}
