% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_param_meta.R
\docType{data}
\name{param_meta}
\alias{param_meta}
\title{Parameter Metadata}
\format{An object of class \code{list} of length 12.}
\usage{
param_meta
}
\description{
Meta data for model parameters including common names, model calls, description and units
}
\author{
Mike Johnson
}
\keyword{datasets}
