% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap.R
\name{read_ftp}
\alias{read_ftp}
\title{Read from FTP}
\usage{
read_ftp(URL, cat, lyrs = 1, AOI, ext = NULL, crs = NULL, dates = NULL)
}
\arguments{
\item{URL}{Unique Resource Identifier (http or local)}

\item{cat}{catalog element}

\item{lyrs}{lyrs to extract}

\item{AOI}{Area of Interest}

\item{ext}{extent of source (if needed)}

\item{crs}{crs of source (if needed)}

\item{dates}{dates of data}
}
\value{
SpatRaster
}
\description{
Read from FTP
}
\seealso{
Other dap: 
\code{\link{.resource_grid}()},
\code{\link{.resource_time}()},
\code{\link{climater_dap}()},
\code{\link{climater_filter}()},
\code{\link{dap}()},
\code{\link{dap_crop}()},
\code{\link{dap_get}()},
\code{\link{dap_meta}()},
\code{\link{dap_summary}()},
\code{\link{dap_to_local}()},
\code{\link{dap_xyzv}()},
\code{\link{extract_sites}()},
\code{\link{get_data}()},
\code{\link{go_get_dap_data}()},
\code{\link{grid_meta}()},
\code{\link{make_ext}()},
\code{\link{make_vect}()},
\code{\link{merge_across_time}()},
\code{\link{parse_date}()},
\code{\link{read_dap_file}()},
\code{\link{time_meta}()},
\code{\link{try_att}()},
\code{\link{var_to_terra}()},
\code{\link{variable_meta}()},
\code{\link{vrt_crop_get}()}
}
\concept{dap}
