% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try_att}
\alias{try_att}
\title{TryCatch around RNetCDF::att.get.nc()}
\usage{
try_att(nc, variable, attribute)
}
\arguments{
\item{nc}{"NetCDF" object which points to the NetCDF dataset. Found with RNetCDF::open.nc.}

\item{variable}{ID or name of the variable from which the attribute will be read, or "NC_GLOBAL" for a global attribute.}

\item{attribute}{Attribute name or ID.}
}
\value{
Vector with a data type that depends on the NetCDF variable. For NetCDF variables of type NC_CHAR, the R type is either character or raw, as specified by argument rawchar. For NC_STRING, the R type is character. Numeric variables are read as double precision by default, but the smallest R type that exactly represents each external type is used if fitnum is TRUE.
}
\description{
TryCatch around RNetCDF::att.get.nc()
}
